robotgo库支持条件编译
